tctl resource: convert ui_config handler
Convert `ui_config` to the new tctl resource handler format that was
introduced in #59518
